+++ src/modules/stateToColumn_mod.f90
@@ -98,6 +98,7 @@ module stateToColumn_mod
   logical :: useFootprintForTovs           ! choose to use a horizontal footprint operator for radiance obs
   logical :: rejectObsNonMonotonicPressure ! choose to reject obs when interpolated column pressure is non-monotonic
   logical :: rejectObsOutsideGlobalGrid    ! choose to reject obs outside a global domain, currently employed for ORCA025 global grid
+  logical :: NNInterpForCloudVars          ! to perform nearest neighbour horizontal interpolation for cloudy variables
 
 contains 
 
@@ -379,6 +380,7 @@ contains
 
     namelist /nams2c/ slantPath_TO_nl, slantPath_TO_tlad, slantPath_RO_nl, slantPath_RA_nl, calcHeightPressIncrOnColumn
     namelist /nams2c/ useFootprintForTovs, rejectObsNonMonotonicPressure, rejectObsOutsideGlobalGrid
+    namelist /nams2c/ NNInterpForCloudVars
 
     write(*,*) 's2c_setupInterpInfo: STARTING'
     call msg_memUsage('s2c_setupInterpInfo')
@@ -403,6 +405,7 @@ contains
       useFootprintForTovs = .false.
       rejectObsNonMonotonicPressure =.true.
       rejectObsOutsideGlobalGrid = .false.
+      NNInterpForCloudVars = .false.
 
       if (.not. utl_isNamelistPresent('NAMS2C','./flnml') ) then
         if ( mmpi_myid == 0 ) then
@@ -2961,7 +2964,8 @@ contains
     real(4) :: lon_deg_r4, lat_deg_r4
     real(4) :: xpos_r4, ypos_r4, xpos2_r4, ypos2_r4
     integer, parameter :: leftIndex = 1, rightIndex = 2, bottomIndex = 1, topIndex = 2
-
+    logical :: isCloudVariable
+    
     numGridpt(:) = 0
 
     lat_deg_r4 = real(interpInfo%stepProcData(procIndex, stepIndex)%allLat(headerIndex, varLevIndex) *  &
@@ -3089,6 +3093,18 @@ contains
         dldy = real(ypos2_r4,8) - real(latIndex,8)
       end if
 
+      if (NNInterpForCloudVars) then
+        if (varLevIndex > 0) then
+          isCloudVariable = vnl_isCloudVar( gsv_getVarNameFromVarLev(stateVector,varLevIndex) )
+        else
+          isCloudVariable = .false.
+        end if
+        if (isCloudVariable) then
+          dldx = real(nint(dldx), 8)
+          dldy = real(nint(dldy), 8)
+        end if
+      end if
+      
       if ( mask(leftIndex ,bottomIndex) ) then
         gridptCount = gridptCount + 1
         latIndexVec(gridptCount) = latIndex